+++ Assets/BeauRoutine/Routine/Internal/Stats.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (C) 2016-2017. Filament Games, LLC. All rights reserved.
+ * Copyright (C) 2016-2018. Filament Games, LLC. All rights reserved.
  * Author:  Alex Beauchesne
  * Date:    3 Apr 2017
  * 
@@ -41,6 +41,8 @@ namespace BeauRoutine.Internal
         WaitEndOfFrame,
         WaitLateUpdate,
         WaitUpdate,
+        WaitThinkUpdate,
+        WaitCustomUpdate,
         Paused
     }