+++ Assets/BeauRoutine/Routine/Internal/YieldPhase.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (C) 2016-2017. Filament Games, LLC. All rights reserved.
+ * Copyright (C) 2016-2018. Filament Games, LLC. All rights reserved.
  * Author:  Alex Beauchesne
  * Date:    1 March 2018
  * 
@@ -19,6 +19,8 @@ namespace BeauRoutine.Internal
         WaitForFixedUpdate,
         WaitForEndOfFrame,
         WaitForLateUpdate,
-        WaitForUpdate
+        WaitForUpdate,
+        WaitForThinkUpdate,
+        WaitForCustomUpdate
     }
 }