+++ Assets/BeauRoutine/Routine/RoutinePhase.cs
@@ -1,5 +1,5 @@
 ﻿/*
- * Copyright (C) 2016-2017. Filament Games, LLC. All rights reserved.
+ * Copyright (C) 2016-2018. Filament Games, LLC. All rights reserved.
  * Author:  Alex Beauchesne
  * Date:    1 March 2018
  * 
@@ -32,6 +32,16 @@ namespace BeauRoutine
         /// <summary>
         /// Routines must be updated manually.
         /// </summary>
-        Manual
+        Manual,
+
+        /// <summary>
+        /// Routines will be updated at a configurable rate.
+        /// </summary>
+        ThinkUpdate,
+
+        /// <summary>
+        /// Routines will be updated at a configurable rate.
+        /// </summary>
+        CustomUpdate,
     }
 }