+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+### Version 0.9.1
+**10 April 2018**
+
+Added support for ThinkUpdate and CustomUpdate phases. Fixed critical bug when yielding a BeauRoutine to an update phase.
+
+#### Features
+
+* Added ``ThinkUpdate`` and ``CustomUpdate`` phases. These will execute at a configurable rate
+* Added ``Paused`` property to ``RoutineIdentity`` to pause all BeauRoutines on a GameObject
+* Added ``RoutineBootstrap`` component. This will configure BeauRoutine settings on Awake
+
+#### Fixes
+
+* Fixed critical issue where BeauRoutines that yielded to a different phase could be skipped during iteration
+* Fixed case where profiling information could be reported inaccurately after a Manual update
+* Fixed potential NullReferenceException if shutting down BeauRoutine during an update phase
+
+#### Improvements
+
+* Script execution order for BeauRoutine is now fixed to 20000, to ensure more consistent timing between projects
+
+#### Breaking Changes
+
+* Added ``RoutineUpdates.SetUpdateRoutineGenerator`` to avoid ambiguous argument types. Previously this was an overload of SetUpdateRoutine, but it has been made into a distinct function.
+
+#### Documentation
+
+* Documented ThinkUpdate and CustomUpdate phases
+* Documented allocations associated with BeauRoutine's debug profiling
+* Corrected spelling errors
+* Corrected an unfinished section
+
 ### Version 0.9.0
 **15 March 2018**